Backend server, ML code
This is the backend server which is a flask code , the app runs through this sever. U have to run to code on separate laptop, the laptop and the mobile on the the app is running should be connected to the same WIFI. Go in settings of the app, edit the IP address. You will get the ip address when u will run the flask code. AFTER connecting it to the IP address, then your app will start performing operations like Object detection , scene detection , text reading.